Add matching options to TypeNameComparer (#12886)
Added matching options (`TypeNameComparison`) for `TypeNames` to better align with what we support in our Nrbf code.

Splits BinaryFormatUtilities and moves the Core only parts to the Core test assembly. Had to move other code and make tweaks to the test project to make this possible.

Starts some tests for CoreNrbfSerializer.

There is a potential race with the desired assembly not being loaded in time. Let `Type.GetType` deal with loading if necessary.